+++ src/components/command/CommandPaletteWrapper.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import { CommandPalette } from './CommandPalette';
+import { useCommandPalette } from '@/contexts/CommandPaletteContext';
+
+export function CommandPaletteWrapper() {
+  const { isOpen, close } = useCommandPalette();
+
+  return <CommandPalette isOpen={isOpen} onClose={close} />;
+}