+++ src/components/command/index.ts
@@ -1,0 +1,6 @@
+export { CommandPalette } from './CommandPalette';
+export { CommandPaletteWrapper } from './CommandPaletteWrapper';
+export { CommandInput } from './CommandInput';
+export { CommandList } from './CommandList';
+export { CommandItem } from './CommandItem';
+export { CommandRegistration } from './CommandRegistration';